safeher-server/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pymongo==4.6.0
pydantic==2.5.0
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
bcrypt==4.1.1
python-multipart==0.0.6
requests==2.31.0
slowapi==0.1.9
apscheduler==3.10.4
python-dotenv==1.0.0
